board API에서 reviseBoard function 분리
