+++ src/lib/postingFunctions.js
@@ -2,7 +2,7 @@ const { boards } = require('../models');
 
 const { boardScheme } = require('./schemes');
 
-const { checkUser } = require('./validations');
+const { checkUser, checkBoard } = require('./validations');
 
 const postFunction = async (req, res, next, type) => {
     try {
@@ -42,6 +42,45 @@ const postFunction = async (req, res, next, type) => {
     }
 };
 
+const reviseFunction = async (req, res, next, type) => {
+    try {
+        const { userId } = req.query;
+        const { boardId } = req.params;
+
+        const { error, value } = boardScheme.validate(req.body);
+        if (error) throw new Error('INVALID_PARAMETERS');
+        const { title, body, level } = value;
+
+        const user = await checkUser(userId);
+        if (!user) throw new Error('INVALID_PARAMETERS');
+
+        const { checkedId, userLevel } = user;
+
+        const board = await checkBoard(boardId);
+        if (!board) throw new Error('INVALID_PARAMETERS');
+
+        const { writerBoardId } = board;
+
+        const isWriterBoard = checkedId === writerBoardId;
+        if (!isWriterBoard) throw new Error('NO_AUTH');
+
+        const writeAuth = level <= userLevel;
+        if (!writeAuth) throw new Error('NO_AUTH');
+
+        const excerpt = body.substring(0, 150);
+
+        await boards.update(
+            { title, body, excerpt, level },
+            { where: { boardId } },
+        );
+
+        /* TODO: 이미지, 파일 정보 수정 (추가, 삭제) */
+        res.json({});
+    } catch (err) {
+        next(err);
+    }
+};
 module.exports = {
     postFunction,
+    reviseFunction,
 };
